test(pg): share the inbox page-size helper so mobile search sees every match
inbox-search's mobile case asserted the seeded complaint was among the results.
It was not — but not because the filter is broken. Probed directly:
_search with that mobileNumber returns 26 rows, all belonging to that citizen,
and 0 rows for an unknown number. The filter is correct.

The suite provisions ONE citizen fixture suite-wide, so by the time this spec
runs that citizen owns dozens of complaints. A mobile matches all of them, the
inbox pages at 10 sorted by `sla ASC` with no total count, and the seeded row
simply is not on page 1. The SRID sibling test passes because one ID matches one
row.

inbox-filters.spec.ts already solved this with a private showAllRows(), and
pgr-inbox-sort-922.spec.ts carried a second copy of it — while inbox-search,
which needs it just as much, had neither. Lifted to utils/employee-ui.ts as
showAllInboxRows (exporting INBOX_SEARCH_RE alongside), repointed both existing
callers at it, and added the call to the mobile case.

This strengthens the assertion rather than relaxing it: it now searches the full
result set instead of whichever ten rows the SLA sort happened to surface.
